moved log buttons to top and added a last play for users to keep track of the state of the game #44
